+++ dill/src/typecast_builder.rs
@@ -41,8 +41,12 @@ where
         self.builder.instance_type_name()
     }
 
-    fn interfaces(&self) -> Vec<InterfaceDesc> {
-        self.builder.interfaces()
+    fn interfaces(&self, clb: &mut dyn FnMut(&InterfaceDesc) -> bool) {
+        self.builder.interfaces(clb);
+    }
+
+    fn metadata<'b, 'c>(&'b self, clb: &'c mut dyn FnMut(&'b dyn std::any::Any) -> bool) {
+        self.builder.metadata(clb)
     }
 
     fn get(&self, cat: &Catalog) -> Result<Arc<dyn Any + Send + Sync>, InjectionError> {
@@ -98,16 +102,66 @@ impl<'a, Iface: 'static + ?Sized> Iterator for TypecastBuilderIterator<'a, Iface
     type Item = TypecastBuilder<'a, Iface>;
     fn next(&mut self) -> Option<Self::Item> {
         if let Some(bindings) = self.bindings {
-            let prev_pos = self.pos;
-            self.pos += 1;
-            bindings.get(prev_pos).map(|b| {
+            if self.pos < bindings.len() {
+                let b = &bindings[self.pos];
+                self.pos += 1;
+
                 // SAFETY: the TypeID key of the `bindings` map is guaranteed to match the
                 // `Iface` type
                 let caster: &TypeCaster<Iface> = b.caster.downcast_ref().unwrap();
-                TypecastBuilder::new(b.builder.as_ref(), caster)
-            })
-        } else {
-            None
+                return Some(TypecastBuilder::new(b.builder.as_ref(), caster));
+            }
         }
+        None
+    }
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
+pub(crate) struct TypecastPredicateBuilderIterator<'a, Iface: 'static + ?Sized, Pred>
+where
+    Pred: Fn(&dyn Builder) -> bool,
+{
+    bindings: Option<&'a Vec<Binding>>,
+    pred: Pred,
+    pos: usize,
+    _dummy: PhantomData<Iface>,
+}
+
+impl<'a, Iface: 'static + ?Sized, Pred> TypecastPredicateBuilderIterator<'a, Iface, Pred>
+where
+    Pred: Fn(&dyn Builder) -> bool,
+{
+    pub(crate) fn new(bindings: Option<&'a Vec<Binding>>, pred: Pred) -> Self {
+        Self {
+            bindings,
+            pred,
+            pos: 0,
+            _dummy: PhantomData,
+        }
+    }
+}
+
+impl<'a, Iface: 'static + ?Sized, Pred> Iterator
+    for TypecastPredicateBuilderIterator<'a, Iface, Pred>
+where
+    Pred: Fn(&dyn Builder) -> bool,
+{
+    type Item = TypecastBuilder<'a, Iface>;
+    fn next(&mut self) -> Option<Self::Item> {
+        if let Some(bindings) = self.bindings {
+            while self.pos < bindings.len() {
+                let b = &bindings[self.pos];
+                self.pos += 1;
+
+                if (self.pred)(b.builder.as_ref()) {
+                    // SAFETY: the TypeID key of the `bindings` map is guaranteed to match the
+                    // `Iface` type
+                    let caster: &TypeCaster<Iface> = b.caster.downcast_ref().unwrap();
+                    return Some(TypecastBuilder::new(b.builder.as_ref(), caster));
+                }
+            }
+        }
+        None
     }
 }