+++ dill/tests/unit.rs
@@ -1,3 +1,4 @@
 #![feature(assert_matches)]
+#![feature(error_generic_member_access)]
 
 mod tests;