+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- It's now possible to associate custom static metadata with builders:
+  ```rust
+  #[component]
+  #[interface(dyn EventHandler)]
+  #[meta(EventHandlerDesc { event_type: "A"})]
+  #[meta(EventHandlerDesc { event_type: "B"})]
+  struct EventHandlerAB;
+  ```
+- New `BuilderExt` trait was added to provide convenient access to metadata and interfaces
+- New `Catalog::builders_for_with_meta()` allows to filter builders by metadata with a custom predicate
+### Changed
+- `Builder::interfaces` method was changed to iteration via callback to avoid allocating a `Vec`
+
 ## [0.8.1] - 2024-05-27
 ### Fixed
 - Fixed pedantic linter warnings